Revert "fix: showing only approved fallen profiles"
