+++ app/all-fallen/FallenList.jsx
@@ -1,15 +1,13 @@
 import Link from "next/link";
-import styles from "./style.module.scss";
-import ProfileCard from "../ProfileCard";
+import styles from "./page.module.scss";
+import ProfileCard from "../components/ProfileCard";
 import StatusMessage from "@/app/components/StatusMessage";
 
 function FallenList({ fallen }) {
-  const approvedFallen = fallen.filter((f) => f.status === "approved");
-
   return (
     <>
-      {approvedFallen.length > 0 ? (
-        approvedFallen.map((fallen) => (
+      {fallen.length > 0 ? (
+        fallen.map((fallen) => (
           <Link
             key={fallen._id}
             className={styles.cardBackground}
