+++ app/all-fallen/page.module.scss
@@ -15,6 +15,12 @@
         font-weight: bold;
         font-size: $text-lg;
     }
+
+    .searchContainer {
+        display: flex;
+        gap: 10px;
+        align-items: center;
+    }
 }
 
 
@@ -35,9 +41,23 @@
     gap: 20px;
     justify-content: center;
     width: 100%;
+
+    .cardBackground {
+        background-color: $elementGradient;
+        border-bottom: 1px solid $lightblueBorder;
+        border-left: 1px solid $lightblueBorder;
+        border-right: 1px solid $lightblueBorder;
+        border-radius: 0px 0px 15px 15px;
+        padding: 10px;
+        width: 250px;
+
+        &:hover {
+            cursor: pointer;
+        }
+    }
 }
 
-@media (max-width: $breakpoint-lg) {
+@media (max-width: 992px) {
     .header {
         text-align: center;
     }
@@ -45,6 +65,16 @@
     .customBubble {
         flex-direction: column !important;
 
+    }
+
+    .searchContainer {
+        width: 100%;
+        flex-direction: column;
+    }
+
+    .searchInput,
+    .searchButton {
+        width: 100%;
     }
 
     .itemsContainer {
