+++ app/components/HobbyDataBubble/index.jsx
@@ -33,6 +33,7 @@ export default function HobbyDataBubble({
         <div className={styles.hobbyContinuers}>
           {sumMode ? hobbyContinuersSum : hobbyContinuers}
         </div>
+        {/* TODO: Add support for both feminine and masculine forms */}
         <div className={styles.hobbyText}>
           {sumMode
             ? ` ממשיכים בדרך של ${fallenName}`
